Move code from deprecated ConditionTranslator to BasicTranslator
